Begin moving code around to test built-route-tree.js
