+++ lib/util/ensure.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function( obj, prop, propDefaultValue ) {
+	if ( obj && typeof obj[ prop ] === 'undefined' ) {
+		obj[ prop ] = propDefaultValue;
+	}
+};
